Add files via upload
The basic structure of this Weather App has been created using the following HTML and CSS code. This structure will contain all the information.Step-by-step we explained our code. To link them all together, we will start by including links for JavaScript and CSS in our HTML. Therefore, the Javascript will be in the body section and the CSS link will be in the head part.